refactor: Score -> Cards countingCard()
