--- src/main/kotlin/blackjack/domain/Constants.kt
@@ -1,6 +1,0 @@
-package blackjack.domain
-
-object Constants {
-    const val BLACKJACK_SCORE = 21
-    const val INITIAL_CARD_COUNT = 2
-}
